requirements.txt
requests 
python-docx 
gradio
huggingface_hub
FastMCP